Extract landing section to its own component
